+++ packages/blade/src/components/Input/BaseInput/BaseInput.tsx
@@ -31,6 +31,25 @@ import { makeSize } from '~utils/makeSize';
 import type { AriaAttributes } from '~utils/makeAccessible';
 import { makeAccessible } from '~utils/makeAccessible';
 
+type CommonAutoCompleteSuggestionTypes =
+  | 'none'
+  | 'name'
+  | 'email'
+  | 'username'
+  | 'password'
+  | 'newPassword'
+  | 'oneTimeCode'
+  | 'telephone'
+  | 'postalCode'
+  | 'countryName'
+  | 'creditCardNumber'
+  | 'creditCardCSC'
+  | 'creditCardExpiry'
+  | 'creditCardExpiryMonth'
+  | 'creditCardExpiryYear';
+
+type WebAutoCompleteSuggestionType = CommonAutoCompleteSuggestionTypes | 'on';
+
 export type BaseInputProps = FormInputLabelProps &
   FormInputValidationProps & {
     /**
@@ -158,32 +177,6 @@ export type BaseInputProps = FormInputLabelProps &
      */
     keyboardReturnKeyType?: 'default' | 'go' | 'done' | 'next' | 'previous' | 'search' | 'send';
     /**
-     * determines what autoComplete suggestion type to show
-     *
-     * Internally it'll render platform specific attributes:
-     *
-     * - web: `autocomplete`
-     * - iOS: `textContentType`
-     * - android: `autoComplete`
-     *
-     */
-    autoCompleteSuggestionType?:
-      | 'none'
-      | 'name'
-      | 'email'
-      | 'username'
-      | 'password'
-      | 'newPassword'
-      | 'oneTimeCode'
-      | 'telephone'
-      | 'postalCode'
-      | 'countryName'
-      | 'creditCardNumber'
-      | 'creditCardCSC'
-      | 'creditCardExpiry'
-      | 'creditCardExpiryMonth'
-      | 'creditCardExpiryYear';
-    /**
      * Element to be rendered on the trailing slot of input field label
      */
     trailingHeaderSlot?: (value?: string) => ReactNode;
@@ -245,18 +238,41 @@ export type BaseInputProps = FormInputLabelProps &
        * The callback function to be invoked when the value of the input field is submitted.
        */
       onSubmit?: FormInputOnEvent;
+      /**
+       * determines what autoComplete suggestion type to show
+       *
+       * Internally it'll render platform specific attributes:
+       *
+       * - web: `autocomplete`
+       * - iOS: `textContentType`
+       * - android: `autoComplete`
+       *
+       */
+      autoCompleteSuggestionType?: CommonAutoCompleteSuggestionTypes;
     };
     web: {
       /**
        * This is a react-native only prop and has no effect on web.
        */
       onSubmit?: undefined;
+      /**
+       * determines what autoComplete suggestion type to show
+       *
+       * Internally it'll render platform specific attributes:
+       *
+       * - web: `autocomplete`
+       * - iOS: `textContentType`
+       * - android: `autoComplete`
+       *
+       */
+      autoCompleteSuggestionType?: WebAutoCompleteSuggestionType;
     };
   }> &
   StyledPropsBlade;
 
 const autoCompleteSuggestionTypeValues = [
   'none',
+  'on',
   'name',
   'email',
   'username',