+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.native.tsx
@@ -8,12 +8,14 @@ import type {
   TouchableHighlightProps,
   GestureResponderEvent,
 } from 'react-native';
+import type { BaseInputProps } from './BaseInput';
 import type { StyledBaseInputProps } from './types';
 import { getBaseInputStyles } from './baseInputStyles';
 import { Text } from '~components/Typography';
 import { assignWithoutSideEffects } from '~utils/assignWithoutSideEffects';
 import { size } from '~tokens/global';
 import { makeSize } from '~utils/makeSize';
+import type { Platform } from '~utils';
 
 type StyledComponentAutoCompleteAndroid =
   | 'off'
@@ -218,7 +220,9 @@ const _StyledBaseInput: React.ForwardRefRenderFunction<
       autoCompleteType={
         autoCompleteSuggestionType
           ? (autoCompleteSuggestionTypeAndroid[
-              autoCompleteSuggestionType
+              autoCompleteSuggestionType as Platform.CastNative<
+                BaseInputProps['autoCompleteSuggestionType']
+              >
             ] as StyledComponentAutoCompleteAndroid)
           : undefined
       }
@@ -226,7 +230,11 @@ const _StyledBaseInput: React.ForwardRefRenderFunction<
       isTextArea={isTextArea}
       textContentType={
         autoCompleteSuggestionType
-          ? autoCompleteSuggestionTypeIOS[autoCompleteSuggestionType]
+          ? autoCompleteSuggestionTypeIOS[
+              autoCompleteSuggestionType as Platform.CastNative<
+                BaseInputProps['autoCompleteSuggestionType']
+              >
+            ]
           : undefined
       }
       autoCapitalize={autoCapitalize}